Move generator utils to a separate class
