+++ shared/src/test/scala/coop/rchain/shared/GeneratorUtils.scala
@@ -1,0 +1,38 @@
+package coop.rchain.shared
+
+import org.scalacheck.Arbitrary.arbitrary
+import org.scalacheck.{Arbitrary, Gen}
+
+object GeneratorUtils {
+
+  /**
+  Credit: https://gist.github.com/etorreborre/d0616e704ed85d7276eb12b025df8ab0
+
+   Distinct list of elements from a given arbitrary
+    */
+  def distinctListOf[T: Arbitrary] =
+    distinctListOfGen(arbitrary[T])(_ == _)
+
+  /**
+  Distinct list of elements from a given generator
+   with a maximum number of elements to discard
+    */
+  def distinctListOfGen[T](gen: Gen[T], maxDiscarded: Int = 1000)(
+      comp: (T, T) => Boolean
+  ): Gen[List[T]] = {
+    val seen      = new scala.collection.mutable.ListBuffer[T]
+    var discarded = 0
+
+    Gen.sized { size =>
+      if (size == seen.size) seen.toList
+      else {
+        while (seen.size <= size && discarded < maxDiscarded) gen.sample match {
+          case Some(t) if !seen.exists(comp(t, _)) =>
+            seen.+=:(t)
+          case _ => discarded += 1
+        }
+        seen.toList
+      }
+    }
+  }
+}
